move functions to correct files
